Type fix

+++ rapid.cpp
@@ -29,10 +29,10 @@ Rapid::Rapid(QString serialConnection) :
     Rapid(serialConnection,this->DEFAULT_RAPID_TYPE,this->DEFAULT_UNLOCK_CODE,this->DEFAULT_VOLTAGE,this->DEFAULT_VIRTUAL_VERSION);
 }
 
-float Rapid::getRapidMinWaitTime(int power, int nPulses, double frequency)
+float Rapid::getRapidMinWaitTime(int power, int nPulses, float frequency)
 {
     float minWaitTime = 0.5;
-    float calcWaitTime = (nPulses * ((frequency * this->JOULES[power]) - 1050.0));
+    float calcWaitTime = (nPulses * ((frequency * this->JOULES[power]) - float(1050.0)));
     if(minWaitTime < calcWaitTime) {
         return calcWaitTime;
     }
@@ -41,7 +41,7 @@ float Rapid::getRapidMinWaitTime(int power, int nPulses, double frequency)
     }
 }
 
-float Rapid::getRapidMaxOnTime(int power, double frequency)
+float Rapid::getRapidMaxOnTime(int power, float frequency)
 {
     float PulseNum = 63000.0;
     float FreqPow = frequency * this->JOULES[power];
